+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Add translations for notes titles and eoc sections to `pl-marketing`
+
 ## [v2.14.0] - 2024-03-22
 
 * Create `BakeExerciseWithTitle` direction for `webview`