+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Create `V2` of `BakeChapterGlossary`
 * Change `note titles` in `information systems`
 * Change EOC sections in `information-systems`